+++ package.json
@@ -92,6 +92,7 @@
     "bn.js": "^4.11.0",
     "create-hash": "^1.1.2",
     "ethereum-cryptography": "^0.1.3",
+    "elliptic": "^6.5.2",
     "ethjs-util": "0.1.6",
     "rlp": "^2.2.3"
   },
