+++ src/secp256k1v3-lib/index.ts
@@ -1,3 +1,6 @@
+// This file is imported from secp256k1 v3
+// https://github.com/cryptocoinjs/secp256k1-node/blob/master/LICENSE
+
 import BN = require('bn.js')
 const EC = require('elliptic').ec
 
@@ -12,7 +15,7 @@ export interface SigObj {
 exports.privateKeyExport = function(privateKey: Buffer, compressed: boolean = true): Buffer {
   const d = new BN(privateKey)
   if (d.ucmp(ecparams.n) >= 0) {
-    throw new Error('private was invalid, try again')
+    throw new Error("couldn't export to DER format")
   }
 
   const point = ec.g.mul(d)
